Fix AMP overflow in style Gram matrix computation
Cast features to fp32 before Gram matrix computation to prevent overflow
when using mixed precision training. Style loss values can exceed fp16 max
(~65,504), causing inf values that corrupt training.

This fix allows AMP to work correctly while maintaining numerical stability.

+++ src/styleforge/train_cin.py
@@ -134,9 +134,11 @@ def train_cin(args):
 
                 content_loss = args.content_weight * mse_loss(features_y.relu2_2, features_x.relu2_2)
 
+                # Compute style loss in fp32 to prevent Gram matrix overflow
                 style_loss = 0.
                 for ft_y, gm_s in zip(features_y, gram_style):
-                    gm_y = utils.gram_matrix(ft_y)
+                    # Cast to fp32 before Gram matrix computation
+                    gm_y = utils.gram_matrix(ft_y.float())
                     style_loss += mse_loss(gm_y, gm_s[:n_batch, :, :])
                 style_loss *= args.style_weight
